+++ validity-bytestring/CHANGELOG.md
@@ -1,2 +1,3 @@
 Unreleased changes:
 * Validity for bytestring is now more precise: It will actually look into the bytestring constructor at the length and the offset.
+* Added validity for lazy bytestrings